Conflicting Next.js configs: remove duplicate config file
