--- next.config.mjs
@@ -1,104 +1,0 @@
-import createNextIntlPlugin from 'next-intl/plugin';
- 
-const withNextIntl = createNextIntlPlugin("./i18n.ts");
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  experimental: {
-    optimizePackageImports: ['lucide-react'],
-  },
-  images: {
-    domains: ['placeholder.svg'],
-    dangerouslyAllowSVG: true,
-    contentSecurityPolicy: "default-src 'self'; script-src 'none'; sandbox;",
-    unoptimized: true,
-  },
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-  // Security headers configuration
-  async headers() {
-    return [
-      {
-        source: '/:path*',
-        headers: [
-          // Content Security Policy (CSP) - Comprehensive protection against XSS and injection attacks
-          {
-            key: 'Content-Security-Policy',
-            value: [
-              "default-src 'self'",
-              "script-src 'self' 'unsafe-eval' 'unsafe-inline'",
-              "style-src 'self' 'unsafe-inline'",
-              "img-src 'self' data: https: blob:",
-              "font-src 'self' data:",
-              "connect-src 'self' https:",
-              "frame-ancestors 'none'",
-              "base-uri 'self'",
-              "form-action 'self'",
-              "object-src 'none'",
-              "upgrade-insecure-requests",
-            ].join('; '),
-          },
-          // HTTP Strict Transport Security (HSTS) - Enforce HTTPS
-          {
-            key: 'Strict-Transport-Security',
-            value: 'max-age=63072000; includeSubDomains; preload',
-          },
-          // Permissions Policy - Control browser features and APIs
-          {
-            key: 'Permissions-Policy',
-            value: [
-              'camera=()',
-              'microphone=()',
-              'geolocation=()',
-              'interest-cohort=()',
-              'payment=()',
-              'usb=()',
-              'bluetooth=()',
-              'magnetometer=()',
-              'gyroscope=()',
-              'accelerometer=()',
-              'ambient-light-sensor=()',
-            ].join(', '),
-          },
-          // X-Frame-Options - Prevent clickjacking (legacy support)
-          {
-            key: 'X-Frame-Options',
-            value: 'DENY',
-          },
-          // X-Content-Type-Options - Prevent MIME type sniffing
-          {
-            key: 'X-Content-Type-Options',
-            value: 'nosniff',
-          },
-          // Referrer-Policy - Control referrer information
-          {
-            key: 'Referrer-Policy',
-            value: 'strict-origin-when-cross-origin',
-          },
-          // X-DNS-Prefetch-Control - Control DNS prefetching
-          {
-            key: 'X-DNS-Prefetch-Control',
-            value: 'on',
-          },
-          // X-XSS-Protection - Legacy XSS protection (for older browsers)
-          {
-            key: 'X-XSS-Protection',
-            value: '1; mode=block',
-          },
-        ],
-      },
-    ]
-  },
-  // Enable static exports for better offline support
-  output: 'standalone',
-  // Optimize for African internet conditions
-  compress: true,
-  poweredByHeader: false,
-  reactStrictMode: true,
-}
-
-export default withNextIntl(nextConfig)
